requirements.txt
motor==2.4.0
Flask==2.0.0
asgiref==3.3.4
Flask-Assets==2.0
python-decouple==3.4
Flask-Discord==0.1.64
python-dotenv==0.17.1